auth_service/requirements.txt
pydantic==2.12.0
pydantic-settings==2.11.0
redis==6.4.0
fastapi[all]==0.111.0
uvicorn==0.38.0
sqlalchemy[asyncio]==2.0.44
alembic==1.17.2
asyncpg==0.30.0
werkzeug==3.1.3
python-jose[cryptography]==3.5.0
typer==0.20.0
flake8==7.3.0